Fix issue with Cucumber driverLifecycle and move integration test to right package
