+++ fluentlenium-cucumber/src/test/java/org/fluentlenium/cucumber/integration/BasicRunner.java
@@ -1,0 +1,12 @@
+package org.fluentlenium.cucumber.integration;
+
+import cucumber.api.CucumberOptions;
+import cucumber.api.junit.Cucumber;
+import org.junit.runner.RunWith;
+
+@RunWith(Cucumber.class)
+@CucumberOptions(features = "classpath:org/fluentlenium/cucumber/integration", format = {"pretty", "html:target/cucumber", "json:target/cucumber.json"})
+public class BasicRunner {
+
+
+}
