+++ fluentlenium-cucumber/src/test/java/org/fluentlenium/cucumber/integration/step/SimpleFeatureMultiStep1.java
@@ -1,19 +1,16 @@
-package org.fluentlenium.cucumber.step;
+package org.fluentlenium.cucumber.integration.step;
 
 import cucumber.api.Scenario;
 import cucumber.api.java.After;
 import cucumber.api.java.Before;
 import cucumber.api.java.en.Given;
-import org.fluentlenium.configuration.ConfigurationProperties;
-import org.fluentlenium.configuration.FluentConfiguration;
 import org.fluentlenium.core.annotation.Page;
 import org.fluentlenium.cucumber.adapter.FluentCucumberTest;
-import org.fluentlenium.cucumber.page.LocalPage;
+import org.fluentlenium.cucumber.integration.page.LocalPage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.htmlunit.HtmlUnitDriver;
 
-@FluentConfiguration(driverLifecycle = ConfigurationProperties.DriverLifecycle.JVM)
-public class SimpleScenarioMultiStep1 extends FluentCucumberTest {
+public class SimpleFeatureMultiStep1 extends FluentCucumberTest {
 
     @Page
     LocalPage page;
@@ -23,7 +20,7 @@ public class SimpleScenarioMultiStep1 extends FluentCucumberTest {
         return new HtmlUnitDriver();
     }
 
-    @Given(value = "scenario multi1 I am on the first page")
+    @Given(value = "feature multi1 I am on the first page")
     public void step1() {
         goTo(page);
     }
