+++ fluentlenium-cucumber/src/test/java/org/fluentlenium/cucumber/integration/step/SimpleFeatureMultiStep2.java
@@ -1,44 +1,32 @@
-package org.fluentlenium.cucumber.step;
+package org.fluentlenium.cucumber.integration.step;
 
 import cucumber.api.Scenario;
 import cucumber.api.java.After;
 import cucumber.api.java.Before;
-import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
-import org.fluentlenium.configuration.ConfigurationProperties;
-import org.fluentlenium.configuration.FluentConfiguration;
 import org.fluentlenium.core.annotation.Page;
 import org.fluentlenium.cucumber.adapter.FluentCucumberTest;
-import org.fluentlenium.cucumber.page.LocalPage;
+import org.fluentlenium.cucumber.integration.page.LocalPage2;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.htmlunit.HtmlUnitDriver;
 
-@FluentConfiguration(driverLifecycle = ConfigurationProperties.DriverLifecycle.JVM)
-public class SimpleFeatureStep extends FluentCucumberTest {
+public class SimpleFeatureMultiStep2 extends FluentCucumberTest {
 
     @Page
-    LocalPage page;
-
-    @Page
-    LocalPage page2;
+    LocalPage2 page2;
 
     @Override
     public WebDriver getDefaultDriver() {
         return new HtmlUnitDriver();
     }
 
-    @Given(value = "feature I am on the first page")
-    public void step1() {
-        goTo(page);
-    }
-
-    @When(value = "feature I click on next page")
+    @When(value = "feature multi2 I click on next page")
     public void step2() {
         $("a#linkToPage2").click();
     }
 
-    @Then(value = "feature I am on the second page")
+    @Then(value = "feature multi2 I am on the second page")
     public void step3() {
         page2.isAt();
     }
@@ -52,4 +40,5 @@ public class SimpleFeatureStep extends FluentCucumberTest {
     public void after(Scenario scenario) {
         super.after(scenario);
     }
+
 }
