+++ fluentlenium-cucumber/src/test/java/org/fluentlenium/cucumber/integration/step/SimpleFeatureStep.java
@@ -1,4 +1,4 @@
-package org.fluentlenium.cucumber.step;
+package org.fluentlenium.cucumber.integration.step;
 
 import cucumber.api.Scenario;
 import cucumber.api.java.After;
@@ -10,12 +10,12 @@ import org.fluentlenium.configuration.ConfigurationProperties;
 import org.fluentlenium.configuration.FluentConfiguration;
 import org.fluentlenium.core.annotation.Page;
 import org.fluentlenium.cucumber.adapter.FluentCucumberTest;
-import org.fluentlenium.cucumber.page.LocalPage;
+import org.fluentlenium.cucumber.integration.page.LocalPage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.htmlunit.HtmlUnitDriver;
 
-@FluentConfiguration(driverLifecycle = ConfigurationProperties.DriverLifecycle.CLASS)
-public class SimpleScenarioStep extends FluentCucumberTest {
+@FluentConfiguration(driverLifecycle = ConfigurationProperties.DriverLifecycle.JVM)
+public class SimpleFeatureStep extends FluentCucumberTest {
 
     @Page
     LocalPage page;
@@ -28,17 +28,17 @@ public class SimpleScenarioStep extends FluentCucumberTest {
         return new HtmlUnitDriver();
     }
 
-    @Given(value = "scenario I am on the first page")
+    @Given(value = "feature I am on the first page")
     public void step1() {
         goTo(page);
     }
 
-    @When(value = "scenario I click on next page")
+    @When(value = "feature I click on next page")
     public void step2() {
         $("a#linkToPage2").click();
     }
 
-    @Then(value = "scenario I am on the second page")
+    @Then(value = "feature I am on the second page")
     public void step3() {
         page2.isAt();
     }
@@ -52,5 +52,4 @@ public class SimpleScenarioStep extends FluentCucumberTest {
     public void after(Scenario scenario) {
         super.after(scenario);
     }
-
 }
