+++ fluentlenium-cucumber/src/test/java/org/fluentlenium/cucumber/integration/step/SimpleScenarioMultiStep2.java
@@ -1,4 +1,4 @@
-package org.fluentlenium.cucumber.step;
+package org.fluentlenium.cucumber.integration.step;
 
 import cucumber.api.Scenario;
 import cucumber.api.java.After;
@@ -9,7 +9,7 @@ import org.fluentlenium.configuration.ConfigurationProperties;
 import org.fluentlenium.configuration.FluentConfiguration;
 import org.fluentlenium.core.annotation.Page;
 import org.fluentlenium.cucumber.adapter.FluentCucumberTest;
-import org.fluentlenium.cucumber.page.LocalPage;
+import org.fluentlenium.cucumber.integration.page.LocalPage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.htmlunit.HtmlUnitDriver;
