+++ fluentlenium-cucumber/src/test/java/org/fluentlenium/cucumber/integration/step/SimpleScenarioStep.java
@@ -1,35 +1,41 @@
-package org.fluentlenium.cucumber.step;
+package org.fluentlenium.cucumber.integration.step;
 
 import cucumber.api.Scenario;
 import cucumber.api.java.After;
 import cucumber.api.java.Before;
+import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
-import org.fluentlenium.configuration.ConfigurationProperties;
-import org.fluentlenium.configuration.FluentConfiguration;
 import org.fluentlenium.core.annotation.Page;
 import org.fluentlenium.cucumber.adapter.FluentCucumberTest;
-import org.fluentlenium.cucumber.page.LocalPage2;
+import org.fluentlenium.cucumber.integration.page.LocalPage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.htmlunit.HtmlUnitDriver;
 
-@FluentConfiguration(driverLifecycle = ConfigurationProperties.DriverLifecycle.METHOD)
-public class SimpleFeatureMultiStep2 extends FluentCucumberTest {
+public class SimpleScenarioStep extends FluentCucumberTest {
 
     @Page
-    LocalPage2 page2;
+    LocalPage page;
+
+    @Page
+    LocalPage page2;
 
     @Override
     public WebDriver getDefaultDriver() {
         return new HtmlUnitDriver();
     }
 
-    @When(value = "feature multi2 I click on next page")
+    @Given(value = "scenario I am on the first page")
+    public void step1() {
+        goTo(page);
+    }
+
+    @When(value = "scenario I click on next page")
     public void step2() {
         $("a#linkToPage2").click();
     }
 
-    @Then(value = "feature multi2 I am on the second page")
+    @Then(value = "scenario I am on the second page")
     public void step3() {
         page2.isAt();
     }
