+++ fluentlenium-cucumber/src/test/java/org/fluentlenium/cucumber/integration/utils/UrlUtil.java
@@ -1,4 +1,4 @@
-package org.fluentlenium.integration.util;
+package org.fluentlenium.cucumber.integration.utils;
 
 import java.net.URL;
